requirements.txt
-i https://pypi.org/simple
annotated-types==0.6.0; python_version >= '3.8'
asgiref==3.7.2; python_version >= '3.7'
click==8.1.7; python_version >= '3.7'
colorama==0.4.6; platform_system == 'Windows'
coloredlogs==15.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django==5.0.2; python_version >= '3.10'
django-ninja==1.1.0; python_version >= '3.7'
flatbuffers==23.5.26
humanfriendly==10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
magika==0.5.0; python_version < '3.12' and python_version >= '3.8'
mpmath==1.3.0
numpy==1.26.4; python_version >= '3.9'
onnxruntime==1.17.0
packaging==23.2; python_version >= '3.7'
protobuf==4.25.3; python_version >= '3.8'
pydantic==2.6.1; python_version >= '3.8'
pydantic-core==2.16.2; python_version >= '3.8'
pyreadline3==3.4.1; python_version >= '3.8' and sys_platform == 'win32'
python-dotenv==1.0.1; python_version >= '3.8'
sqlparse==0.4.4; python_version >= '3.5'
sympy==1.12; python_version >= '3.8'
tabulate==0.9.0; python_version >= '3.7'
tqdm==4.66.2; python_version >= '3.7'
typing-extensions==4.9.0; python_version >= '3.8'
tzdata==2024.1; sys_platform == 'win32'
